добавяет деление на модули
